optimization fork
purpose: many http requests (heh) and content loaded inefficiently

local css is now in style.css with minification of frameworks.css,
font-mfizz.css, and unminified dogecoin.css

mfizz font moved to more logical folder name. references updated and
stylesheet inlined.

font-awesome script removed from the header as it is [near] identical
to the font awesome that has been imbedded in frameworks.css

images run through pngcrush to make them ever so much smaller. every
little bit counts.

add this and translate triple the page load time so they’re moved to
the bottom of the page for efficient loading times and also it lets
dogecoin.js and remodel load as soon as the visual bit of the page is
loading.