restored exchanges section, added a link to price comparison
I propose to add a link https://exchangerates.pro/?to_currency=DOGE that will prepare all the possible ways to buy DOGE in the user country and local currency, if possible. Our platform helps the user to find the most profitable and safe way to buy, sell, and exchange crypto in 255 countries comparing Exchanges, Brokers and the P2P Market. We believe it will be useful. Thanks!